[feature/pickling] Splitting up the bindings
